+++ src/pyCRT/arrayPlotting.py
@@ -26,7 +26,7 @@ from .curveFitting import (
 # Type aliases for commonly used types
 # {{{
 # Array of arbitraty size with float elements.
-Array = NDArray[np.float_]
+Array = NDArray[np.float64]
 
 # Tuples of two numpy arrays, typically an array of the timestamp for each
 # frame and an array of average intensities within a given ROI
@@ -40,10 +40,10 @@ ParameterSequence = Union[Sequence[float], Array]
 # optimized parameters and the second their standard deviations
 FitParametersTuple = tuple[ParameterSequence, ParameterSequence]
 
-Real = Union[float, int, np.float_, np.int_]
+Real = Union[float, int, np.float64, int]
 
-# This accounts for the fact that np.int_ doesn't inherit from int
-Integer = Union[int, np.int_]
+# This accounts for the fact that int doesn't inherit from int
+Integer = Union[int, int]
 
 FigAxTuple = tuple[Figure, Axes]
 # }}}
